requirements.txt
streamlit
pandas
numpy
pyresparser
pdfminer.six
nltk
spacy
plotly
Pillow
pymysql
streamlit-tags
pafy
yt-dlp
geocoder
geopy